+++ framework/base/Formatter.php
@@ -38,15 +38,15 @@ class Formatter extends Component
     /**
      * @var string the default format string to be used to format a date using PHP date() function.
      */
-    public $dateFormat = 'Y/m/d';
+    public $dateFormat = 'Y-m-d';
     /**
      * @var string the default format string to be used to format a time using PHP date() function.
      */
-    public $timeFormat = 'h:i:s A';
+    public $timeFormat = 'H:i:s';
     /**
      * @var string the default format string to be used to format a date and time using PHP date() function.
      */
-    public $datetimeFormat = 'Y/m/d h:i:s A';
+    public $datetimeFormat = 'Y-m-d H:i:s';
     /**
      * @var string the text to be displayed when formatting a null. Defaults to '<span class="not-set">(not set)</span>'.
      */